+++ API/src/controllers/admin.js
@@ -25,7 +25,7 @@ const changeBan = async (req, res) => {
   try {
     const user = await User.findByPk(id);
     await user.update({
-        isBanned: !user.isBanned,
+      isBanned: !user.isBanned,
     });
     res.json(user);
   } catch (error) {
@@ -49,8 +49,24 @@ const getCategories = async (req, res) => {
   }
 };
 
+const changeRole = async (req, res) => {
+  const { id } = req.params;
+  try {
+    const user = await User.findByPk(id);
+    const roleId = user.roleAdminId;
+    let role = await RoleAdmin.findByPk(roleId);
+    role.name =
+      role.name === "ADMIN" ? (role.name = "USER") : (role.name = "ADMIN");
+    await role.update({ name: role.name });
+    res.send("Successful update");
+  } catch (error) {
+    res.status(404).json({ error: error.message });
+  }
+};
+
 module.exports = {
   getUsers,
   changeBan,
   getCategories,
+  changeRole,
 };
